I think I got backpropagation working, tested on a simple function and seemed to work well

+++ include/cost.hh
@@ -2,4 +2,4 @@
 
 #include <cmath>
 
-float quadraticCost(double expected, double actual, bool prime);
+double quadraticCost(double expected, double actual, bool prime);
