+++ src/cost.cc
@@ -1,5 +1,5 @@
 #include "cost.hh"
 
-float quadraticCost(double expected, double actual, bool prime) {
-    return prime ? actual - expected : 0.5 * std::pow(expected - actual, 2.0);
+double quadraticCost(double expected, double actual, bool prime) {
+    return prime ? (actual - expected) : (0.5 * std::pow(expected - actual, 2.0));
 }
